Try to prevent virtual keyboards to pop up in mobile
This would happen when a user touches the canvas - or the invisible div
above it to be more precise, which accepts text input and especially paste
events. We try to enable pasting images from the clipboard by having this
div in place.
If we find touch events to be present, we completely(?) disable this feature
for now, since on iOS and Android clipboards aren't supposed to contain
images anyway. On other devices (windows tablets) this feature is now
unavailable until someone finds a better solution.

Fix #36