[I18n] Initial implementation
This commit adds a language picker component as well as a Localization-
Manager class which keeps a list of managed DOM elements to be translated
when the user changes the language. Only two locales (en_US and de_DE) have
been added so far (json files) and they only contain localized strings for
three small text components, so a lot of work still needs to be done. Also,
dynamically adding new dom nodes to be localized by the LocalizationManager
has not been used anywhere yet.

This work is done to eventually implement #79