[Letters Effect] Reduce the amount of generated shader code
We were running into shader size limitations, so we decided to change
the whole rendering approach. We now use a texture to store 8 line segments
for each particle which we can access in the fragment shader. This
reduces shader code immensely and also gives us the possibility to change
fonts from within js