Try to not respond to changes of devicePixelRatio when zooming
I don't know why browser zooming would cause resize events, but on FF mobile
it seems to happen sporadically. With a messed-up DPR value on zoom things
would therefore be weird. Now we just cache the initial value and always
use that. Let's see what happens when we do that...