[Recording] Try patching WebMRecorder for real-time recording
A video I recorded on my smartphone would play the content in high-speed
only to settle on the final frame for the remainder of the time the video
should actually run for. This is (probably) caused by the fact that
'addFrame' jobs queue up because of non-realtime encoding speeds. We now
repeat frames as long as it takes for a new frame to be encoded. While this
is cheating wrt the actual frame data in the video, it places matching
frames at the correct positions in the video. So if your hardware is too
slow for real-time encoding, you will get a slide-show - but at least it's
at the right speed